[rrc] split the back leg into driver::backend
backend(), backend_module(), and unit_output() — everything that takes
the front leg's product through the MLIR pipeline and LLVM to the
requested artifact — move together.
